requirements.txt
langchain
langchainhub
langchain_community
gradio
gradio_client
tavily-python
faiss-cpu
beautifulsoup4
pandas
numpy
langchain-core
langchain-openai
langchain-text-splitters
langgraph
google-generativeai
google-ai-generativelanguage
langchain-google-genai
langchain-tavily
python-dotenv
yfinance
cachetools
streamlit
plotly